moved logic to handling env vars for wrapped MET config files into util/met_config
