+++ internal_tests/pytests/command_builder/test_command_builder.py
@@ -381,19 +381,6 @@ def test_handle_description(metplus_config, config_overrides, expected_value):
     assert cbw.env_var_dict.get('METPLUS_DESC', '') == expected_value
 
 @pytest.mark.parametrize(
-    'input, output', [
-        ('', 'NONE'),
-        ('NONE', 'NONE'),
-        ('FCST', 'FCST'),
-        ('OBS', 'OBS'),
-        ('G002', '"G002"'),
-    ]
-)
-def test_format_regrid_to_grid(metplus_config, input, output):
-    cbw = CommandBuilder(metplus_config())
-    assert cbw.format_regrid_to_grid(input) == output
-
-@pytest.mark.parametrize(
     'config_overrides, set_to_grid, expected_dict', [
         ({}, True, {'REGRID_TO_GRID': 'NONE'}),
         ({}, False, {}),
@@ -499,25 +486,29 @@ def test_handle_regrid_new(metplus_config, config_overrides, expected_output):
          True, 'test_string_1 = value_1;'),
     ]
 )
-def test_set_met_config_string(metplus_config, mp_config_name, met_config_name,
+def test_add_met_config_string(metplus_config, mp_config_name, met_config_name,
                                c_dict_key, remove_quotes, expected_output):
     cbw = CommandBuilder(metplus_config())
 
     # set some config variables to test
     cbw.config.set('config', 'TEST_STRING_1', 'value_1')
 
-    c_dict = {}
+    extra_args = {}
+    if remove_quotes:
+        extra_args['remove_quotes'] = True
 
-    cbw.set_met_config_string(c_dict,
-                              mp_config_name,
-                              met_config_name,
-                              c_dict_key=c_dict_key,
-                              remove_quotes=remove_quotes)
     key = c_dict_key
     if key is None:
-        key = met_config_name.upper()
+        key = met_config_name
+    key = key.upper()
 
-    assert c_dict.get(key, '') == expected_output
+    cbw.add_met_config(name=met_config_name,
+                       data_type='string',
+                       env_var_name=key,
+                       metplus_configs=[mp_config_name],
+                       extra_args=extra_args)
+
+    assert cbw.env_var_dict.get(f'METPLUS_{key}', '') == expected_output
 
 @pytest.mark.parametrize(
     'mp_config_name,met_config_name,c_dict_key,uppercase,expected_output, is_ok', [
@@ -547,7 +538,7 @@ def test_set_met_config_string(metplus_config, mp_config_name, met_config_name,
          True, '', False),
     ]
 )
-def test_set_met_config_bool(metplus_config, mp_config_name, met_config_name,
+def test_add_met_config_bool(metplus_config, mp_config_name, met_config_name,
                              c_dict_key, uppercase, expected_output, is_ok):
     cbw = CommandBuilder(metplus_config())
 
@@ -556,18 +547,22 @@ def test_set_met_config_bool(metplus_config, mp_config_name, met_config_name,
     cbw.config.set('config', 'TEST_BOOL_3', False)
     cbw.config.set('config', 'TEST_BOOL_4', 'chicken')
 
-    c_dict = {}
+    extra_args = {}
+    if not uppercase:
+        extra_args['uppercase'] = False
 
-    cbw.set_met_config_bool(c_dict,
-                            mp_config_name,
-                            met_config_name,
-                            c_dict_key=c_dict_key,
-                            uppercase=uppercase)
     key = c_dict_key
     if key is None:
-        key = met_config_name.upper()
+        key = met_config_name
+    key = key.upper()
 
-    assert c_dict.get(key, '') == expected_output
+    cbw.add_met_config(name=met_config_name,
+                       data_type='bool',
+                       env_var_name=key,
+                       metplus_configs=[mp_config_name],
+                       extra_args=extra_args)
+
+    assert cbw.env_var_dict.get(f'METPLUS_{key}', '') == expected_output
     assert cbw.isOK == is_ok
 
 # int
@@ -590,7 +585,7 @@ def test_set_met_config_bool(metplus_config, mp_config_name, met_config_name,
          '', False),
     ]
 )
-def test_set_met_config_int(metplus_config, mp_config_name, met_config_name,
+def test_add_met_config_int(metplus_config, mp_config_name, met_config_name,
                              c_dict_key, expected_output, is_ok):
     cbw = CommandBuilder(metplus_config())
 
@@ -599,17 +594,17 @@ def test_set_met_config_int(metplus_config, mp_config_name, met_config_name,
     cbw.config.set('config', 'TEST_INT_3', -4)
     cbw.config.set('config', 'TEST_INT_4', 'chicken')
 
-    c_dict = {}
-
-    cbw.set_met_config_int(c_dict,
-                           mp_config_name,
-                           met_config_name,
-                           c_dict_key=c_dict_key)
     key = c_dict_key
     if key is None:
-        key = met_config_name.upper()
+        key = met_config_name
+    key = key.upper()
 
-    assert c_dict.get(key, '') == expected_output
+    cbw.add_met_config(name=met_config_name,
+                       data_type='int',
+                       env_var_name=key,
+                       metplus_configs=[mp_config_name])
+
+    assert cbw.env_var_dict.get(f'METPLUS_{key}', '') == expected_output
     assert cbw.isOK == is_ok
 
 @pytest.mark.parametrize(
@@ -631,7 +626,7 @@ def test_set_met_config_int(metplus_config, mp_config_name, met_config_name,
          '', False),
     ]
 )
-def test_set_met_config_float(metplus_config, mp_config_name, met_config_name,
+def test_add_met_config_float(metplus_config, mp_config_name, met_config_name,
                              c_dict_key, expected_output, is_ok):
     cbw = CommandBuilder(metplus_config())
 
@@ -640,17 +635,17 @@ def test_set_met_config_float(metplus_config, mp_config_name, met_config_name,
     cbw.config.set('config', 'TEST_FLOAT_3', 4)
     cbw.config.set('config', 'TEST_FLOAT_4', 'chicken')
 
-    c_dict = {}
-
-    cbw.set_met_config_float(c_dict,
-                             mp_config_name,
-                             met_config_name,
-                             c_dict_key=c_dict_key)
     key = c_dict_key
     if key is None:
-        key = met_config_name.upper()
+        key = met_config_name
+    key = key.upper()
 
-    assert c_dict.get(key, '') == expected_output
+    cbw.add_met_config(name=met_config_name,
+                       data_type='float',
+                       env_var_name=key,
+                       metplus_configs=[mp_config_name])
+
+    assert cbw.env_var_dict.get(f'METPLUS_{key}', '') == expected_output
     assert cbw.isOK == is_ok
 
 @pytest.mark.parametrize(
@@ -678,7 +673,7 @@ def test_set_met_config_float(metplus_config, mp_config_name, met_config_name,
          'test_thresh_6 = NA;', True),
     ]
 )
-def test_set_met_config_thresh(metplus_config, mp_config_name, met_config_name,
+def test_add_met_config_thresh(metplus_config, mp_config_name, met_config_name,
                                c_dict_key, expected_output, is_ok):
     cbw = CommandBuilder(metplus_config())
 
@@ -689,17 +684,18 @@ def test_set_met_config_thresh(metplus_config, mp_config_name, met_config_name,
     cbw.config.set('config', 'TEST_THRESH_5', '>CDP40&&<=CDP50')
     cbw.config.set('config', 'TEST_THRESH_6', 'NA')
 
-    c_dict = {}
-
-    cbw.set_met_config_thresh(c_dict,
-                              mp_config_name,
-                              met_config_name,
-                              c_dict_key=c_dict_key)
     key = c_dict_key
     if key is None:
-        key = met_config_name.upper()
+        key = met_config_name
+    key = key.upper()
 
-    assert c_dict.get(key, '') == expected_output
+    cbw.add_met_config(name=met_config_name,
+                       env_var_name=key,
+                       data_type='thresh',
+                       metplus_configs=[mp_config_name])
+
+    print(f"KEY: {key}, ENV VARS: {cbw.env_var_dict}")
+    assert cbw.env_var_dict.get(f'METPLUS_{key}', '') == expected_output
     assert cbw.isOK == is_ok
 
 @pytest.mark.parametrize(
@@ -727,7 +723,7 @@ def test_set_met_config_thresh(metplus_config, mp_config_name, met_config_name,
          True, 'test_list_4 = [value_1, value2];'),
     ]
 )
-def test_set_met_config_list(metplus_config, mp_config_name, met_config_name,
+def test_add_met_config_list(metplus_config, mp_config_name, met_config_name,
                              c_dict_key, remove_quotes, expected_output):
     cbw = CommandBuilder(metplus_config())
 
@@ -736,18 +732,23 @@ def test_set_met_config_list(metplus_config, mp_config_name, met_config_name,
     cbw.config.set('config', 'TEST_LIST_3', "'value_1',   'value2'")
     cbw.config.set('config', 'TEST_LIST_4', '"value_1",   "value2"')
 
-    c_dict = {}
+    extra_args = {}
+    if remove_quotes:
+        extra_args['remove_quotes'] = True
 
-    cbw.set_met_config_list(c_dict,
-                            mp_config_name,
-                            met_config_name,
-                            c_dict_key=c_dict_key,
-                            remove_quotes=remove_quotes)
     key = c_dict_key
     if key is None:
-        key = met_config_name.upper()
+        key = met_config_name
 
-    assert c_dict.get(key, '') == expected_output
+    key = key.upper()
+
+    cbw.add_met_config(name=met_config_name,
+                       data_type='list',
+                       env_var_name=key,
+                       metplus_configs=[mp_config_name],
+                       extra_args=extra_args)
+    print(f"KEY: {key}, ENV VARS: {cbw.env_var_dict}")
+    assert cbw.env_var_dict.get(f'METPLUS_{key}', '') == expected_output
 
 @pytest.mark.parametrize(
     'mp_config_name,allow_empty,expected_output', [
@@ -761,40 +762,26 @@ def test_set_met_config_list(metplus_config, mp_config_name, met_config_name,
         ('TEST_LIST_2', True, ''),
     ]
 )
-def test_set_met_config_list_allow_empty(metplus_config, mp_config_name,
+def test_add_met_config_list_allow_empty(metplus_config, mp_config_name,
                                          allow_empty, expected_output):
     cbw = CommandBuilder(metplus_config())
 
     # set some config variables to test
     cbw.config.set('config', 'TEST_LIST_1', '')
 
-    c_dict = {}
+    extra_args = {}
+    if allow_empty:
+        extra_args['allow_empty'] = True
 
     met_config_name = mp_config_name.lower()
 
-    cbw.set_met_config_list(c_dict,
-                            mp_config_name,
-                            met_config_name,
-                            allow_empty=allow_empty)
+    cbw.add_met_config(name=met_config_name,
+                       data_type='list',
+                       metplus_configs=[mp_config_name],
+                       extra_args=extra_args)
 
-    assert c_dict.get(mp_config_name, '') == expected_output
-
-@pytest.mark.parametrize(
-    'data_type, expected_function', [
-        ('int', 'set_met_config_int'),
-        ('float', 'set_met_config_float'),
-        ('list', 'set_met_config_list'),
-        ('string', 'set_met_config_string'),
-        ('thresh', 'set_met_config_thresh'),
-        ('bool', 'set_met_config_bool'),
-        ('bad_name', None),
-    ]
-)
-def test_set_met_config_function(metplus_config, data_type, expected_function):
-    cbw = CommandBuilder(metplus_config())
-    function_found = cbw.set_met_config_function(data_type)
-    function_name = function_found.__name__ if function_found else None
-    assert(function_name == expected_function)
+    assert cbw.env_var_dict.get(f'METPLUS_{mp_config_name}', '') == expected_output
+    #assert c_dict.get(mp_config_name, '') == expected_output
 
 def test_add_met_config_dict(metplus_config):
     dict_name = 'fcst_hr_window'
