+++ metplus/wrappers/command_builder.py
@@ -24,7 +24,8 @@ from ..util import config_metplus
 from ..util import METConfig
 from ..util import MISSING_DATA_VALUE
 from ..util import get_custom_string_list
-from ..util import get_wrapped_met_config_file
+from ..util import get_wrapped_met_config_file, add_met_config_item, format_met_config
+from ..util.met_config import format_regrid_to_grid
 
 # pylint:disable=pointless-string-statement
 '''!@namespace CommandBuilder
@@ -281,20 +282,6 @@ class CommandBuilder:
             env_var_value = do_string_sub(raw_env_var_value,
                                           **time_info)
             self.add_env_var(env_var, env_var_value)
-
-    @staticmethod
-    def format_regrid_to_grid(to_grid):
-        to_grid = to_grid.strip('"')
-        if not to_grid:
-            to_grid = 'NONE'
-
-        # if NONE, FCST, or OBS force uppercase, otherwise add quotes
-        if to_grid.upper() in ['NONE', 'FCST', 'OBS']:
-            to_grid = to_grid.upper()
-        else:
-            to_grid = f'"{to_grid}"'
-
-        return to_grid
 
     def print_all_envs(self, print_copyable=True, print_each_item=True):
         """! Create list of log messages that output all environment variables
@@ -1450,297 +1437,6 @@ class CommandBuilder:
 
         return time_info
 
-    def _get_config_or_default(self, mp_config_name, get_function,
-                               default=None):
-        conf_value = ''
-
-        # if no possible METplus config variables are not set
-        if mp_config_name is None:
-            # if no default defined, return without doing anything
-            if not default:
-                return None
-
-        # if METplus config variable is set, read the value
-        else:
-            conf_value = get_function('config',
-                                      mp_config_name,
-                                      '')
-
-        # if variable is not set and there is a default defined, set default
-        if not conf_value and default:
-            conf_value = default
-
-        return conf_value
-
-    def set_met_config_list(self, c_dict, mp_config, met_config_name,
-                            c_dict_key=None, **kwargs):
-        """! Get list from METplus configuration file and format it to be passed
-              into a MET configuration file. Set c_dict item with formatted string.
-             Args:
-                 @param c_dict configuration dictionary to set
-                 @param mp_config_name METplus configuration variable name. Assumed to be
-                  in the [config] section. Value can be a comma-separated list of items.
-                 @param met_config name of MET configuration variable to set. Also used
-                  to determine the key in c_dict to set (upper-case)
-                 @param c_dict_key optional argument to specify c_dict key to store result. If
-                  set to None (default) then use upper-case of met_config_name
-                 @param allow_empty if True, if METplus config variable is set
-                  but is an empty string, then set the c_dict value to an empty
-                  list. If False, behavior is the same as when the variable is
-                  not set at all, which is to not set anything for the c_dict
-                  value
-                 @param remove_quotes if True, output value without quotes.
-                  Default value is False
-                 @param default (Optional) if set, use this value as default
-                  if config is not set
-        """
-        mp_config_name = self.get_mp_config_name(mp_config)
-        conf_value = self._get_config_or_default(
-            mp_config_name,
-            get_function=self.config.getraw,
-            default=kwargs.get('default')
-        )
-        if conf_value is None:
-            return
-
-        # convert value from config to a list
-        conf_values = util.getlist(conf_value)
-        if conf_values or kwargs.get('allow_empty', False):
-            out_values = []
-            for conf_value in conf_values:
-                remove_quotes = kwargs.get('remove_quotes', False)
-                # if not removing quotes, escape any quotes found in list items
-                if not remove_quotes:
-                    conf_value = conf_value.replace('"', '\\"')
-
-                conf_value = util.remove_quotes(conf_value)
-                if not remove_quotes:
-                    conf_value = f'"{conf_value}"'
-
-                out_values.append(conf_value)
-            out_value = f"[{', '.join(out_values)}]"
-
-            if not c_dict_key:
-                c_key = met_config_name.upper()
-            else:
-                c_key = c_dict_key
-
-            conf_value = f'{met_config_name} = {out_value};'
-            c_dict[c_key] = conf_value
-
-    def set_met_config_string(self, c_dict, mp_config, met_config_name,
-                              c_dict_key=None, **kwargs):
-        """! Get string from METplus configuration file and format it to be passed
-              into a MET configuration file. Set c_dict item with formatted string.
-
-                 @param c_dict configuration dictionary to set
-                 @param mp_config METplus configuration variable name. Assumed to be
-                  in the [config] section. Value can be a comma-separated list of items.
-                 @param met_config_name name of MET configuration variable to set. Also used
-                  to determine the key in c_dict to set (upper-case)
-                 @param c_dict_key optional argument to specify c_dict key to store result. If
-                  set to None (default) then use upper-case of met_config_name
-                 @param remove_quotes if True, output value without quotes.
-                  Default value is False
-                 @param to_grid if True, format to_grid value
-                  Default value is False
-                 @param default (Optional) if set, use this value as default
-                  if config is not set
-        """
-        mp_config_name = self.get_mp_config_name(mp_config)
-        conf_value = self._get_config_or_default(
-            mp_config_name,
-            get_function=self.config.getraw,
-            default=kwargs.get('default')
-        )
-        if not conf_value:
-            return
-
-        conf_value = util.remove_quotes(conf_value)
-        # add quotes back if remote quotes is False
-        if not kwargs.get('remove_quotes'):
-            conf_value = f'"{conf_value}"'
-
-        if kwargs.get('uppercase', False):
-            conf_value = conf_value.upper()
-
-        if kwargs.get('to_grid', False):
-            conf_value = self.format_regrid_to_grid(conf_value)
-
-        c_key = c_dict_key if c_dict_key else met_config_name.upper()
-        c_dict[c_key] = f'{met_config_name} = {conf_value};'
-
-    def set_met_config_number(self, c_dict, num_type, mp_config,
-                              met_config_name, c_dict_key=None, **kwargs):
-        """! Get integer from METplus configuration file and format it to be passed
-              into a MET configuration file. Set c_dict item with formatted string.
-             Args:
-                 @param c_dict configuration dictionary to set
-                 @param num_type type of number to get from config. If set to 'int', call
-                   getint function. If not, call getfloat function.
-                 @param mp_config METplus configuration variable name. Assumed to be
-                  in the [config] section. Value can be a comma-separated list of items.
-                 @param met_config_name name of MET configuration variable to set. Also used
-                  to determine the key in c_dict to set (upper-case) if c_dict_key is None
-                 @param c_dict_key optional argument to specify c_dict key to store result. If
-                  set to None (default) then use upper-case of met_config_name
-                 @param default (Optional) if set, use this value as default
-                  if config is not set
-        """
-        mp_config_name = self.get_mp_config_name(mp_config)
-        if mp_config_name is None:
-            return
-
-        if num_type == 'int':
-            conf_value = self.config.getint('config', mp_config_name)
-        else:
-            conf_value = self.config.getfloat('config', mp_config_name)
-
-        if conf_value is None:
-            self.isOK = False
-        elif conf_value != util.MISSING_DATA_VALUE:
-            if not c_dict_key:
-                c_key = met_config_name.upper()
-            else:
-                c_key = c_dict_key
-
-            c_dict[c_key] = f"{met_config_name} = {str(conf_value)};"
-
-    def set_met_config_int(self, c_dict, mp_config_name, met_config_name,
-                           c_dict_key=None, **kwargs):
-        self.set_met_config_number(c_dict, 'int',
-                                   mp_config_name,
-                                   met_config_name,
-                                   c_dict_key=c_dict_key,
-                                   **kwargs)
-
-    def set_met_config_float(self, c_dict, mp_config_name,
-                             met_config_name, c_dict_key=None, **kwargs):
-        self.set_met_config_number(c_dict, 'float',
-                                   mp_config_name,
-                                   met_config_name,
-                                   c_dict_key=c_dict_key,
-                                   **kwargs)
-
-    def set_met_config_thresh(self, c_dict, mp_config, met_config_name,
-                              c_dict_key=None, **kwargs):
-        mp_config_name = self.get_mp_config_name(mp_config)
-        if mp_config_name is None:
-            return
-
-        conf_value = self.config.getstr('config', mp_config_name, '')
-        if conf_value:
-            if util.get_threshold_via_regex(conf_value) is None:
-                self.log_error(f"Incorrectly formatted threshold: {mp_config_name}")
-                return
-
-            if not c_dict_key:
-                c_key = met_config_name.upper()
-            else:
-                c_key = c_dict_key
-
-            c_dict[c_key] = f"{met_config_name} = {str(conf_value)};"
-
-    def set_met_config_bool(self, c_dict, mp_config, met_config_name,
-                            c_dict_key=None, **kwargs):
-        """! Get boolean from METplus configuration file and format it to be
-             passed into a MET configuration file. Set c_dict item with boolean
-             value expressed as a string.
-             Args:
-                 @param c_dict configuration dictionary to set
-                 @param mp_config METplus configuration variable name.
-                  Assumed to be in the [config] section.
-                 @param met_config_name name of MET configuration variable to
-                  set. Also used to determine the key in c_dict to set
-                  (upper-case)
-                 @param c_dict_key optional argument to specify c_dict key to
-                  store result. If set to None (default) then use upper-case of
-                  met_config_name
-                 @param uppercase If true, set value to TRUE or FALSE
-        """
-        mp_config_name = self.get_mp_config_name(mp_config)
-        if mp_config_name is None:
-            return
-        conf_value = self.config.getbool('config', mp_config_name, '')
-        if conf_value is None:
-            self.log_error(f'Invalid boolean value set for {mp_config_name}')
-            return
-
-        # if not invalid but unset, return without setting c_dict with no error
-        if conf_value == '':
-            return
-
-        conf_value = str(conf_value)
-        if kwargs.get('uppercase', True):
-            conf_value = conf_value.upper()
-
-        if not c_dict_key:
-            c_key = met_config_name.upper()
-        else:
-            c_key = c_dict_key
-
-        c_dict[c_key] = (f'{met_config_name} = '
-                         f'{util.remove_quotes(conf_value)};')
-
-    def get_mp_config_name(self, mp_config):
-        """! Get first name of METplus config variable that is set.
-
-        @param mp_config list of METplus config keys to check. Can also be a
-        single item
-        @returns Name of first METplus config name in list that is set in the
-        METplusConfig object. None if none keys in the list are set.
-        """
-        if not isinstance(mp_config, list):
-            mp_configs = [mp_config]
-        else:
-            mp_configs = mp_config
-
-        for mp_config_name in mp_configs:
-            if self.config.has_option('config', mp_config_name):
-                return mp_config_name
-
-        return None
-
-    @staticmethod
-    def format_met_config(data_type, c_dict, name, keys=None):
-        """! Return formatted variable named <name> with any <items> if they
-        are set to a value. If none of the items are set, return empty string
-
-        @param data_type type of value to format
-        @param c_dict config dictionary to read values from
-        @param name name of dictionary to create
-        @param keys list of c_dict keys to use if they are set. If unset (None)
-         then read all keys from c_dict
-        @returns MET config formatted dictionary/list
-         if any items are set, or empty string if not
-        """
-        values = []
-        if keys is None:
-            keys = c_dict.keys()
-
-        for key in keys:
-            value = c_dict.get(key)
-            if value:
-                values.append(str(value))
-
-        # if none of the keys are set to a value in dict, return empty string
-        if not values:
-            return ''
-
-        output = ''.join(values)
-        # add curly braces if dictionary
-        if 'dict' in data_type:
-            output = f"{{{output}}}"
-
-        # add square braces if list
-        if 'list' in data_type:
-            output = f"[{output}];"
-
-        # if name is not empty, add variable name and equals sign
-        if name:
-            output = f'{name} = {output}'
-        return output
-
     @staticmethod
     def format_met_config_dict(c_dict, name, keys=None):
         """! Return formatted dictionary named <name> with any <items> if they
@@ -1753,59 +1449,28 @@ class CommandBuilder:
         @returns MET config formatted dictionary if any items are set, or empty
          string if not
         """
-        return CommandBuilder.format_met_config('dict', c_dict=c_dict, name=name, keys=keys)
+        return format_met_config('dict', c_dict=c_dict, name=name, keys=keys)
 
     def handle_regrid(self, c_dict, set_to_grid=True):
-        app_name_upper = self.app_name.upper()
-
-        # dictionary to hold regrid values as they are read
-        tmp_dict = {}
-
+        dict_items = {}
         if set_to_grid:
+            dict_items['to_grid'] = ('string', 'to_grid')
+
+            # handle legacy format of to_grid
             conf_value = (
                 self.config.getstr('config',
-                                   f'{app_name_upper}_REGRID_TO_GRID', '')
+                                   f'{self.app_name.upper()}_REGRID_TO_GRID', '')
             )
 
             # set to_grid without formatting for backwards compatibility
-            formatted_to_grid = self.format_regrid_to_grid(conf_value)
+            formatted_to_grid = format_regrid_to_grid(conf_value)
             c_dict['REGRID_TO_GRID'] = formatted_to_grid
 
-            if conf_value:
-                tmp_dict['REGRID_TO_GRID'] = (
-                    f"to_grid = {formatted_to_grid};"
-                )
-
-        self.set_met_config_string(tmp_dict,
-                                   f'{app_name_upper}_REGRID_METHOD',
-                                   'method',
-                                   c_dict_key='REGRID_METHOD',
-                                   remove_quotes=True)
-
-        self.set_met_config_int(tmp_dict,
-                                f'{app_name_upper}_REGRID_WIDTH',
-                                'width',
-                                c_dict_key='REGRID_WIDTH')
-
-        self.set_met_config_float(tmp_dict,
-                                  f'{app_name_upper}_REGRID_VLD_THRESH',
-                                  'vld_thresh',
-                                  c_dict_key='REGRID_VLD_THRESH')
-        self.set_met_config_string(tmp_dict,
-                                   f'{app_name_upper}_REGRID_SHAPE',
-                                   'shape',
-                                   c_dict_key='REGRID_SHAPE',
-                                   remove_quotes=True)
-
-        regrid_string = self.format_met_config_dict(tmp_dict,
-                                                    'regrid',
-                                                    ['REGRID_TO_GRID',
-                                                     'REGRID_METHOD',
-                                                     'REGRID_WIDTH',
-                                                     'REGRID_VLD_THRESH',
-                                                     'REGRID_SHAPE',
-                                                     ])
-        self.env_var_dict['METPLUS_REGRID_DICT'] = regrid_string
+        dict_items['method'] = ('string', 'uppercase,remove_quotes')
+        dict_items['width'] = 'int'
+        dict_items['vld_thresh'] = 'float'
+        dict_items['shape'] = ('string', 'uppercase,remove_quotes')
+        self.add_met_config_dict('regrid', dict_items)
 
     def handle_description(self):
         """! Get description from config. If <app_name>_DESC is set, use
@@ -2123,82 +1788,6 @@ class CommandBuilder:
 
         self.add_met_config_dict('mask', items)
 
-    def set_met_config_function(self, item_type):
-        """! Return function to use based on item type
-
-             @param item_type type of MET config variable to obtain
-             Valid values: list, string, int, float, thresh, bool
-             @returns function to use or None if invalid type provided
-        """
-        if item_type == 'int':
-            return self.set_met_config_int
-        elif item_type == 'string':
-            return self.set_met_config_string
-        elif item_type == 'list':
-            return self.set_met_config_list
-        elif item_type == 'float':
-            return self.set_met_config_float
-        elif item_type == 'thresh':
-            return self.set_met_config_thresh
-        elif item_type == 'bool':
-            return self.set_met_config_bool
-        else:
-            self.log_error("Invalid argument for item type: "
-                           f"{item_type}")
-            return None
-
-    def add_met_config_item(self, item, output_dict=None, depth=0):
-        """! Reads info from METConfig object, gets value from
-        METplusConfig, and formats it based on the specifications. Sets
-        value in output dictionary with key starting with METPLUS_.
-
-        @param item METConfig object to read and determine what to get
-        @param output_dict (optional) dictionary to save formatted output
-         If unset, use self.env_var_dict.
-        @param depth counter to check if item being processed is nested within
-         another variable or not. If depth is 0, it is a top level variable.
-         This is used internally by this function and shouldn't be supplied
-         outside of calls within this function.
-        """
-        if output_dict is None:
-            output_dict = self.env_var_dict
-
-        env_var_name = item.env_var_name.upper()
-        if not env_var_name.startswith('METPLUS_'):
-            env_var_name = f'METPLUS_{env_var_name}'
-
-        # handle dictionary or dictionary list item
-        if 'dict' in item.data_type:
-            tmp_dict = {}
-            for child in item.children:
-                if not self.add_met_config_item(child, tmp_dict,
-                                                   depth=depth+1):
-                    return False
-
-            dict_string = self.format_met_config(item.data_type,
-                                                 tmp_dict,
-                                                 item.name,
-                                                 keys=None)
-
-            # if handling dict MET config that is not nested inside another
-            if not depth and item.data_type == 'dict':
-                env_var_name = f'{env_var_name}_DICT'
-
-            output_dict[env_var_name] = dict_string
-            return True
-
-        # handle non-dictionary item
-        set_met_config = self.set_met_config_function(item.data_type)
-        if not set_met_config:
-            return False
-
-        set_met_config(output_dict,
-                       item.metplus_configs,
-                       item.name,
-                       c_dict_key=env_var_name,
-                       **item.extra_args)
-        return True
-
     def add_met_config_dict(self, dict_name, items):
         """! Read config variables for MET config dictionary and set
          env_var_dict with formatted values
@@ -2278,7 +1867,13 @@ class CommandBuilder:
             children=dict_items,
         )
 
-        return self.add_met_config_item(final_met_config, self.env_var_dict)
+        return_code = add_met_config_item(self.config,
+                                          final_met_config,
+                                          self.env_var_dict)
+        if not return_code:
+            self.isOK = False
+
+        return return_code
 
     @staticmethod
     def parse_item_info(item_info):
@@ -2345,8 +1940,9 @@ class CommandBuilder:
                 f"{self.app_name}_{kwargs.get('name')}".upper()
             ]
         item = METConfig(**kwargs)
-        output_dict = kwargs.get('output_dict')
-        self.add_met_config_item(item, output_dict)
+        output_dict = kwargs.get('output_dict', self.env_var_dict)
+        if not add_met_config_item(self.config, item, output_dict):
+            self.isOK = False
 
     def get_met_config(self, **kwargs):
         """! Get METConfig object from arguments and return it
